インスタンス、コード簡略化
