+++ DREAM CATCH/YoutubeViewController.swift
@@ -27,11 +27,9 @@ class YoutubeViewController: UIViewController {
     }
     
     private func fetchYoutubeSerachInfo() {
-        let path = "search"
         let params = ["q": "lebronjames"]
         
-        apiRequest(path: path, params: params, type: Video.self) { (video) in
-            
+        APIRequest.shared.request(path: .search, params: params, type: Video.self) { (video) in
             self.videoItems = video.items
             let id = self.videoItems[0].snippet.channelId
             self.fetchYoutubeChannelInfo(id: id)
@@ -39,41 +37,16 @@ class YoutubeViewController: UIViewController {
     }
     
     private func fetchYoutubeChannelInfo(id: String) {
-        let path = "channels"
         let params = [
             "id": id
         ]
         
-        apiRequest(path: path, params: params, type: Channel.self) { (channel) in
+        APIRequest.shared.request(path: .channels, params: params, type: Channel.self) { (channel) in
             self.videoItems.forEach { (item) in
                 item.channel = channel
             }
             
             self.videoListCollectionView.reloadData()
-        }
-    }
-    
-    private func apiRequest<T: Decodable>(path: String, params: [String: Any], type: T.Type, completion: @escaping (T) -> Void) {
-        let baseUrl = "https://www.googleapis.com/youtube/v3/"
-        let path = path
-        let url = baseUrl + path + "?"
-        
-        var params = params
-        params["key"] = "AIzaSyBMchvJKgYkEVGgw2BdAcT9q_-y6pu9QHA"
-        params["part"] = "snippet"
-        
-        let request = AF.request(url, method: .get, parameters: params)
-        
-        request.responseJSON { (response) in
-            do {
-                guard let data = response.data else { return }
-                let decode = JSONDecoder()
-                let value = try decode.decode(T.self, from: data)
-                
-                completion(value)
-            } catch {
-                print("変換に失敗しました。: ", error)
-            }
         }
     }
